+++ app/components/finance/ecf/contract_band_row.rb
@@ -3,7 +3,7 @@
 module Finance
   module ECF
     class ContractBandRow < BaseComponent
-      include ECFPaymentsHelper
+      include FinanceHelper
 
     private
